adding more methods

+++ contracts/IStakeableNFT.sol
@@ -9,6 +9,6 @@ interface IStakeableNFT {
     function tokenURIHash(uint256 tokenId) external view returns (string memory);
     function tokenContributionAsset(uint256 tokenId) external view returns (address);
     function tokenContributionSize(uint256 tokenId) external view returns (uint256);
-    function tokenContributionMultiplier(uint256 tokenId) external view returns (uint256);
+    function tokenContributionMultiplier(uint256 tokenId) external view returns (uint8);
 
 }
